pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Difference generator for configuration files"
authors = ["Pavel Krivosheev <pkrivosheev@bk.ru>"]
readme = "README.md"
packages = [
    { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
